Extract tests for isWeakMap to own file
